Remove access

+++ app/src/main/java/co/dito/abako/apijitpack/utils/backupDocument/BackupDocumentImp.kt
@@ -1,5 +1,6 @@
 package co.dito.abako.apijitpack.utils.backupDocument
 
+import android.content.Context
 import android.os.Environment
 import com.google.gson.Gson
 import java.io.BufferedReader
@@ -9,15 +10,12 @@ import java.io.FileOutputStream
 import java.io.InputStreamReader
 import java.io.OutputStreamWriter
 
-class BackupDocumentImp<T : BackupRequestData> : BackupDocument<T> {
+class BackupDocumentImp<T : BackupRequestData>(context: Context) : BackupDocument<T> {
 
-    private val backupFolder = File("${Environment.getExternalStorageDirectory().absolutePath}/Abako/Backup/")
+    private val backupFolder = context.getExternalFilesDir(Environment.DIRECTORY_DOCUMENTS)
     private lateinit var backupFileDocument: File
 
     override fun getBackup(backupName: String, tClass: Class<T>): BackupRequestData? {
-        if (!backupFolder.exists() && !backupFolder.mkdirs()) {
-            throw Exception("Error al crear la carpeta de backup")
-        }
         backupFileDocument = File(backupFolder, backupName)
         if (!backupFileDocument.exists()) {
             return null
@@ -44,6 +42,10 @@ class BackupDocumentImp<T : BackupRequestData> : BackupDocument<T> {
     }
 
     override fun deleteFiles() {
+        if (backupFolder == null) {
+            return
+        }
+
         if (backupFolder.isDirectory) {
             backupFolder.listFiles()?.forEach { file ->
                 file.delete()
